final check-in
input layer --   384 = 128x3
LSTM layer  --  67584 = 128x128x(3 + 1 label) + 16?
pitch dense --  16512 =     128x(128+1 label)
duration dense -- 129 =       1x(128+1 label)
step dense  --    129 =       1x(128+1 label)
67584 + 16512 + 129 + 129 = 84354 trainable parameters

From the full clean_midi dataset I was able to harvest over 706k notes to the input training set. To reduce the loss I use weighted sum of the 3 component losses for `pitch`, `step` and `duration`. In the generated note distribution plots by pitch-step-duration, it is noticeable the slight change in distribution of the original note variables. Since there is a feedback loop between the model's outputs and inputs, the model tends to generate similar sequences of outputs to reduce the loss. This is particularly relevant for `step` and `duration`, which uses the MSE loss. To increase the randomness of the `pitch` increase the `temperature` in `predict_next_note`.

Overall this was a good project for me. I learned a lot about music, and if not composing it myself from scratch, to at least what statistics lead to a good music learning model and pleasant to the ear tunes. I gained some insight knowledge of `keras` and `tensrflow`. While I am a strong rock-and-roll and hard-rock affectionado, I noticed that training a simple RNN/LSTM `music generation` model on rock and pop pieces is very simplistic as far as pitch, step, duration histograms, nowhere near close to normal distribution which lead to awful music generation pieces based on those music genres. My last attempt (checked-in here) is based on all 363 pieces. I was curious why the 10000 note RNN/LSTM generated musing is plying in a matter of 1-2-3 seconds. The answer in my opinion is the `step` histogram is very heavy around 0.0 sec, which implies that the generated sequential notes with setp=0.0 are played instanteniusly. To that regard at first I tried to pick only music pieces with `step` normally distributed for training, but it was a lot of work looking at so many distributions. Then I came up with a NUDGE factor, currently at 0.025 sec, or 25 ms. This would hopefully ensure that notes with 0.0-based step will get assigned a small, non-zero step, just to nudge those notes from playing simultaneously, thus having a generated music piece that is a bit longer than 1-2-3 seconds.

The current model still has pretty high loss, and maybe I will re-train it with higher numbers of epochs and patience. Also, the idea to nudge the 0.0 step to the arbitrary 25 ms didn't quite workout with this training set, as it can be seen from the generated step histogram all generated music has that step.